Safari-611.1.10/LayoutTests/fast/line-grid/line-grid-parsing-expected.txt

Test parsing of the CSS line-grid property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test("-webkit-line-grid: none") is "none"
PASS test("-webkit-line-grid: first-grid") is "first-grid"
PASS test("-webkit-line-grid: 'first grid'") is ""
PASS test("-webkit-line-grid: ;") is ""
PASS test("-webkit-line-grid: 1") is ""
PASS test("-webkit-line-grid: 1.2") is ""
PASS test("-webkit-line-grid: -1") is ""
PASS test("-webkit-line-grid: 12px") is ""
PASS testComputedStyle("none") is "none"
PASS testComputedStyle("") is "none"
PASS testComputedStyle("'first-flow'") is "none"
PASS testComputedStyle("first-grid") is "first-grid"
PASS testComputedStyle("12px") is "none"
PASS testInherited("none") is "none"
PASS testInherited("grid") is "grid"
PASS successfullyParsed is true

TEST COMPLETE
